Warlock/erc1155

+++ contracts/mocks/ERC721Mock.sol
@@ -9,15 +9,15 @@ contract ERC721Mock is ERC721 {
     constructor(string memory _name, string memory _symbol) ERC721(_name, _symbol) {}
     string public baseTokenURI;
 
-    function mint(address to, uint256 tokenId) public {
+    function mint(address to, uint tokenId) public {
         _safeMint(to, tokenId, "");
     }
 
-    function transfer(address to, uint256 tokenId) public {
+    function transfer(address to, uint tokenId) public {
         _safeTransfer(msg.sender, to, tokenId, "");
     }
 
-    function isApprovedOrOwner(address spender, uint256 tokenId) public view virtual returns (bool) {
+    function isApprovedOrOwner(address spender, uint tokenId) public view virtual returns (bool) {
         return _isApprovedOrOwner(spender, tokenId);
     }
 }
